Move increment package name to their own file
