+++ x-pack/plugins/fleet/server/services/package_policy.test.ts
@@ -49,8 +49,6 @@ import {
   updatePackageInputs,
   packagePolicyService,
   _applyIndexPrivileges,
-  incrementPackageName,
-  incrementPackagePolicyCopyName,
 } from './package_policy';
 import { appContextService } from './app_context';
 import { fetchInfo } from './epm/registry';
@@ -3242,85 +3240,5 @@ describe('_applyIndexPrivileges()', () => {
 
     const streamOut = _applyIndexPrivileges(packageStream, inputStream);
     expect(streamOut).toEqual(expectedStream);
-  });
-
-  describe('increment package name', () => {
-    it('should return 1 if no existing policies', async () => {
-      packagePolicyService.list = jest.fn().mockResolvedValue(undefined);
-      const newName = await incrementPackageName(savedObjectsClientMock.create(), 'apache');
-      expect(newName).toEqual('apache-1');
-    });
-
-    it('should return 11 if max policy name is 10', async () => {
-      packagePolicyService.list = jest.fn().mockResolvedValue({
-        items: [
-          { name: 'apache-1' },
-          { name: 'aws-11' },
-          { name: 'apache-10' },
-          { name: 'apache-9' },
-        ],
-      });
-      const newName = await incrementPackageName(savedObjectsClientMock.create(), 'apache');
-      expect(newName).toEqual('apache-11');
-    });
-  });
-
-  describe('increment package policy copy name', () => {
-    it('should return packagePolicyName (copy) if no existing policies', async () => {
-      packagePolicyService.list = jest.fn().mockResolvedValue({ items: [] });
-      const newName = await incrementPackagePolicyCopyName(
-        savedObjectsClientMock.create(),
-        'packagePolicyName'
-      );
-      expect(newName).toEqual('packagePolicyName (copy)');
-    });
-
-    it('should return packagePolicyName (copy 2) if there is an existing copy', async () => {
-      packagePolicyService.list = jest.fn().mockResolvedValue({
-        items: [
-          {
-            name: 'packagePolicyName (copy)',
-          },
-        ],
-      });
-      const newName = await incrementPackagePolicyCopyName(
-        savedObjectsClientMock.create(),
-        'packagePolicyName'
-      );
-      expect(newName).toEqual('packagePolicyName (copy 2)');
-    });
-
-    it('should return packagePolicyName (copy 2) if there is an existing copy and copying a copy', async () => {
-      packagePolicyService.list = jest.fn().mockResolvedValue({
-        items: [
-          {
-            name: 'packagePolicyName (copy)',
-          },
-        ],
-      });
-      const newName = await incrementPackagePolicyCopyName(
-        savedObjectsClientMock.create(),
-        'packagePolicyName (copy)'
-      );
-      expect(newName).toEqual('packagePolicyName (copy 2)');
-    });
-
-    it('should return packagePolicyName (copy 3) if there is 2 copy', async () => {
-      packagePolicyService.list = jest.fn().mockResolvedValue({
-        items: [
-          {
-            name: 'packagePolicyName (copy)',
-          },
-          {
-            name: 'packagePolicyName (copy 2)',
-          },
-        ],
-      });
-      const newName = await incrementPackagePolicyCopyName(
-        savedObjectsClientMock.create(),
-        'packagePolicyName'
-      );
-      expect(newName).toEqual('packagePolicyName (copy 3)');
-    });
   });
 });
